pyproject.toml
    [build-system]
    requires = ["setuptools>=61.0.0", "wheel"]
    build-backend = "setuptools.build_meta"

    [project]
    name = "Alexandria"
    version = "0.1.0"
    dependencies = [
        "numpy",
        "langchain",
        "langchain-google-genai",
        "langchain_community",
        "unstructured[md,pdf]",
        "google-generativeai",
        "python-dotenv",
        "pypdf",
        "chromadb",
    ]

[tool.setuptools.packages.find]
where = ["alexandria"]
